olive-cli added - remove snpe workflow, add command files - add main also - olive-cli console_script
