+++ olive/command/__main__.py
@@ -2,7 +2,10 @@
 # Copyright (c) Microsoft Corporation. All rights reserved.
 # Licensed under the MIT License.
 # --------------------------------------------------------------------------
-from olive.workflows.snpe.convertquantize.convertquantize import convertquantize
-from olive.workflows.snpe.evaluate.evaluate import evaluate
+# This is to support running Olive CLI as a module in case olive-cli command
+# is not available in the PATH.
+# Example: python -m olive.command
+from olive.command.olive_cli import main
 
-__all__ = ["convertquantize", "evaluate"]
+if __name__ == "__main__":
+    main()
