+++ olive/command/base.py
@@ -1,0 +1,20 @@
+# -------------------------------------------------------------------------
+# Copyright (c) Microsoft Corporation. All rights reserved.
+# Licensed under the MIT License.
+# --------------------------------------------------------------------------
+from abc import ABC, abstractmethod
+from argparse import ArgumentParser, Namespace
+
+
+class BaseOliveCLICommand(ABC):
+    def __init__(self, args: Namespace):
+        self.args = args
+
+    @staticmethod
+    @abstractmethod
+    def register_subcommand(parser: ArgumentParser):
+        raise NotImplementedError
+
+    @abstractmethod
+    def run(self):
+        raise NotImplementedError
