+++ olive/workflows/__init__.py
@@ -2,6 +2,6 @@
 # Copyright (c) Microsoft Corporation. All rights reserved.
 # Licensed under the MIT License.
 # --------------------------------------------------------------------------
-from olive.workflows.run.run import run
+from olive.workflows.run.run import add_run_args, run
 
-__all__ = ["run"]
+__all__ = ["add_run_args", "run"]
