+++ olive/workflows/run/run.py
@@ -265,6 +265,23 @@ def run(
         return run_engine(package_config, run_config, data_root)
 
 
+def add_run_args(parser):
+    parser.add_argument(
+        "--package-config",
+        type=str,
+        required=False,
+        help=(
+            "For advanced users. Path to optional package (json) config file with location "
+            "of individual pass module implementation and corresponding dependencies."
+            "Configuration might also include user owned/proprietary/private pass implementations."
+        ),
+    )
+    parser.add_argument("--run-config", "--config", type=str, help="Path to json config file", required=True)
+    parser.add_argument("--setup", help="Whether run environment setup", action="store_true")
+    parser.add_argument("--data-root", "--data_root", help="The data root path for optimization", required=False)
+    parser.add_argument("--tempdir", type=str, help="Root directory for tempfile directories and files", required=False)
+
+
 def check_local_ort_installation(package_name: str):
     """Check whether ORT is installed. If not, will return current package name to install."""
     local_ort_packages = get_local_ort_packages()
